dev/feature: implementshield benefit and fix mainMenu and load levels

+++ src/js/loaders/collisionloader.js
@@ -2,9 +2,10 @@ import Phaser from 'phaser-ce';
 import currentGameState from '../currentGameState';
 import { enemyExplode, bossExplode } from '../sound/explosures';
 import { Benefit } from '../objects/benefits';
+import config from '../config';
 
 function randBenefit(){
-  return game.rnd.integerInRange(1, 2);
+  return game.rnd.integerInRange(1, 3);
 }
 
 function invokeSound(that, target) {
@@ -20,11 +21,11 @@ function killEnemies(bullet, enemy) {
   let enemY = enemy.body.center.y;
   enemy.kill();
   invokeSound(this, 'enemy');
-  bullet.kill();
+  if(bullet != this.mainPlayerShield) bullet.kill();
   currentGameState.score += 100;
   currentGameState.levelscore += 100;
   //------------------------benefit health----------------------------
-  if(!this.benefitHealth && randBenefit() == 1 && !this.benefitScore){
+  if(!this.benefitHealth && !this.benefitScore && !this.benefitShield && randBenefit() == 1){
     this.benefitHealth = new Benefit({
       game: this,
       x: enemX,
@@ -33,12 +34,8 @@ function killEnemies(bullet, enemy) {
     });
     this.game.add.existing(this.benefitHealth);
   }
-  //------------------------benefit ammo----------------------------
-
-  //------------------------benefit shield----------------------------
-
   //------------------------benefit score----------------------------
-  if(!this.benefitScore && randBenefit() == 2 && !this.benefitHealth){
+  if(!this.benefitHealth && !this.benefitScore && !this.benefitShield && randBenefit() == 2){
     this.benefitScore = new Benefit({
       game: this,
       x: enemX,
@@ -47,6 +44,19 @@ function killEnemies(bullet, enemy) {
     });
     this.game.add.existing(this.benefitScore);
   }
+
+  //------------------------benefit shield----------------------------
+  if(!this.benefitHealth && !this.benefitScore && !this.benefitShield && !this.mainPlayerShield && randBenefit() == 3){
+    this.benefitShield = new Benefit({
+      game: this,
+      x: enemX,
+      y: enemY,
+      asset: 'shield',
+    });
+    this.game.add.existing(this.benefitShield);
+  }
+
+  //------------------------benefit ammo----------------------------
 }
 
 function killBoss(boss, bullet) {
@@ -54,8 +64,9 @@ function killBoss(boss, bullet) {
   if(boss.HP)boss.HP--;
   if(boss.HP == 0){
     boss.kill();
+
     invokeSound(this, 'boss');
-    boss.endLevel = this.time.now;
+    this.countdown = this.time.now;
     currentGameState.bosskilled = true;
     this.bossWeapon.bullets.destroy();
     currentGameState.score += 1000;
@@ -66,9 +77,10 @@ function killBoss(boss, bullet) {
 function overlapEnemies(player, enemy) {
   enemy.kill();
   if(player.HP)player.HP--;
-  
   if(!player.HP){
     player.kill();
+    currentGameState.mainPlayerKilled = true;
+    this.countdown = this.time.now;
   }
   currentGameState.score += 100;
   currentGameState.levelscore += 100;
@@ -76,11 +88,15 @@ function overlapEnemies(player, enemy) {
 
 function killPlayer(player, bullet) {
   bullet.kill();
-  invokeSound(this, 'enemy');
-  player.HP--;
-  if(!player.HP){
-    player.kill();
-  }
+  //invokeSound(this, 'enemy');//----------------------------------need to fix----------------
+  if(player != this.mainPlayerShield){
+    if(player.HP)player.HP--;
+    if(!player.HP){
+      player.kill();
+      currentGameState.mainPlayerKilled = true;
+      this.countdown = this.time.now;
+    }
+  }  
 }
 
 
@@ -89,17 +105,32 @@ export default function () {
   this.physics.arcade.overlap(this.weapon.bullets, this.boss, killBoss, null, this);
   this.physics.arcade.overlap(this.mainPlayer, this.enemies, overlapEnemies, null, this);
   if(this.bossWeapon){
-    this.physics.arcade.collide(this.bossWeapon.bullets, this.mainPlayer, killPlayer);
+    this.physics.arcade.overlap(this.bossWeapon.bullets, this.mainPlayer, killPlayer, null, this);
   }  
   //-------------------------------Benefits Collisions----------------------------------------------------------  
   if(this.benefitHealth){
-    this.physics.arcade.collide(this.mainPlayer, this.benefitHealth, this.benefitHealth.getHealth, null, this);
+    this.physics.arcade.overlap(this.mainPlayer, this.benefitHealth, this.benefitHealth.getHealth, null, this);
     if(this.benefitHealth && this.benefitHealth.x < 0)this.benefitHealth = null;
   }
 
   if(this.benefitScore){
-    this.physics.arcade.collide(this.mainPlayer, this.benefitScore, this.benefitScore.getScore, null, this);
+    this.physics.arcade.overlap(this.mainPlayer, this.benefitScore, this.benefitScore.getScore, null, this);
     if(this.benefitScore && this.benefitScore.x < 0)this.benefitScore = null;
+  }
+  if(this.benefitShield){
+    this.physics.arcade.overlap(this.mainPlayer, this.benefitShield, this.benefitShield.getShield, null, this);
+    if(this.benefitShield && this.benefitShield.x < 0)this.benefitShield = null;
+  }
+  //---------------------------------------------shieldOn---------------------------------------------------------
+  if(this.mainPlayerShield){
+    this.mainPlayerShield.x = this.mainPlayer.x;
+    this.mainPlayerShield.y = this.mainPlayer.y;
+    this.physics.arcade.overlap(this.mainPlayerShield, this.enemies, killEnemies, null, this);
+    if(this.bossWeapon)this.physics.arcade.overlap(this.bossWeapon.bullets, this.mainPlayerShield, killPlayer, null, this);
+    if(this.time.now > this.mainPlayerShield.countdown + config.shieldDuration || currentGameState.bosskilled){
+      this.mainPlayerShield.kill();
+      this.mainPlayerShield = null;
+    } 
   }
 }
